back-end/dev-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile dev-requirements.in
#
astroid==3.0.1
    # via pylint
black==23.10.1
    # via -r dev-requirements.in
build==1.0.3
    # via pip-tools
click==8.1.7
    # via
    #   black
    #   pip-tools
coverage==7.3.2
    # via -r dev-requirements.in
dill==0.3.7
    # via pylint
isort==5.12.0
    # via pylint
mccabe==0.7.0
    # via pylint
mypy-extensions==1.0.0
    # via black
packaging==23.2
    # via
    #   black
    #   build
pathspec==0.11.2
    # via black
pip-tools==7.3.0
    # via -r dev-requirements.in
platformdirs==3.11.0
    # via
    #   black
    #   pylint
pylint==3.0.2
    # via
    #   pylint-django
    #   pylint-plugin-utils
pylint-django==2.5.5
    # via -r dev-requirements.in
pylint-plugin-utils==0.8.2
    # via pylint-django
pyproject-hooks==1.0.0
    # via build
tomlkit==0.12.2
    # via pylint
wheel==0.41.3
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
